# sdks/go/api/openapi.yaml
openapi: 3.1.0
info:
  description: "Agent credential broker. Use injects the secret at the edge. Responses\
    \ never include vault material. CLI, MCP, and generated SDKs are projections of\
    \ this contract."
  title: Vortex Password Manager
  version: 0.1.0
servers:
- url: https://pwm.vortex.nyc
security:
- bearerAuth: []
tags:
- description: Bearer is the agent. listItems is metadata. useItem injects. Never
    GetSecret.
  name: Agent
paths:
  /health:
    get:
      operationId: getHealth
      responses:
        "200":
          content:
            text/plain:
              schema:
                type: string
          description: ok
      security: []
      summary: Liveness. No secrets.
      tags:
      - Agent
  /openapi.json:
    get:
      operationId: getOpenApi
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
          description: OpenAPI document
      security: []
      summary: This contract.
      tags:
      - Agent
  /v1/items:
    get:
      operationId: listItems
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemsResponse"
          description: Granted items
        "401":
          description: missing or invalid Bearer
      summary: Items this agent may Use. Names and URIs. Never secrets.
      tags:
      - Agent
  /v1/use:
    post:
      operationId: useItem
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UseRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UseResponse"
          description: "Decision plus upstream result, scrubbed"
        "400":
          description: bad request
        "401":
          description: missing or invalid Bearer
      summary: Call a URL as this agent. The broker injects the credential. The vault
        secret is never in the response.
      tags:
      - Agent
components:
  schemas:
    Item:
      additionalProperties: false
      example:
        id: id
        org_id: org_id
        name: name
        kind: api_key
        owner:
          kind: kind
          id: id
        uris:
        - uris
        - uris
        tags:
        - tags
        - tags
        archived: true
        has_totp: true
        has_file: true
      properties:
        id:
          type: string
        org_id:
          type: string
        name:
          type: string
        kind:
          enum:
          - api_key
          - oauth
          - ssh
          - file
          type: string
        owner:
          $ref: "#/components/schemas/Owner"
        uris:
          items:
            type: string
          type: array
          default: null
        tags:
          items:
            type: string
          type: array
          default: null
        archived:
          type: boolean
        has_totp:
          type: boolean
        has_file:
          type: boolean
      required:
      - id
      - kind
      - name
      - org_id
      - owner
      - uris
    Owner:
      additionalProperties: false
      example:
        kind: kind
        id: id
      properties:
        kind:
          type: string
        id:
          type: string
      required:
      - id
      - kind
    ItemsResponse:
      additionalProperties: false
      example:
        items:
        - id: id
          org_id: org_id
          name: name
          kind: api_key
          owner:
            kind: kind
            id: id
          uris:
          - uris
          - uris
          tags:
          - tags
          - tags
          archived: true
          has_totp: true
          has_file: true
        - id: id
          org_id: org_id
          name: name
          kind: api_key
          owner:
            kind: kind
            id: id
          uris:
          - uris
          - uris
          tags:
          - tags
          - tags
          archived: true
          has_totp: true
          has_file: true
      properties:
        items:
          items:
            $ref: "#/components/schemas/Item"
          type: array
          default: null
      required:
      - items
    UseRequest:
      additionalProperties: false
      example:
        item: item
        url: url
        method: GET
        headers:
          key: headers
        body: body
      properties:
        item:
          description: Granted item name
          type: string
        url:
          description: Absolute URL. Host must match the item.
          type: string
        method:
          default: GET
          type: string
        headers:
          additionalProperties:
            type: string
          description: Extra request headers. Never the vault secret.
        body:
          description: Request body. Never the vault secret. Use --body-file on the
            CLI.
          type: string
      required:
      - item
      - url
    UseResponse:
      additionalProperties: false
      example:
        decision: allow
        reason: reason
        approval_id: approval_id
        status: 0
        body: body
      properties:
        decision:
          enum:
          - allow
          - deny
          - need_approval
          type: string
        reason:
          type: string
        approval_id:
          type: string
        status:
          type: integer
        body:
          description: Upstream body with vault secrets scrubbed.
          type: string
      required:
      - decision
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: Hydra client_credentials JWT or bound OIDC. Identity is the agent.
        The host is not identity.
      scheme: bearer
      type: http